Use the most efficient Selector implementation available on the current platform

Odoo supports only SelectSelector but it is a little obsolete

python >= 3.4 supports a new high-level library Selectors:
 - https://docs.python.org/es/3/library/selectors.html

It could to auto-choose the following ones:
 - SelectSelector
 - PollSelector
 - EpollSelector
 - DevpollSelector
 - KqueueSelector

Using the DefaultSelector class the most efficient implementation available on the current platform will be use:
 - https://docs.python.org/3/library/selectors.html#selectors.DefaultSelector

It helps to support better the resources of the system

Using SelectSelector you are not able to run workers >=255

If you set `ulimit -n 10240` and run `odoo-bin --workers=255`
the following error is raised:

    Traceback (most recent call last):
    File "odoo/service/server.py", line 926, in run
        self.sleep()
    File "odoo/service/server.py", line 852, in sleep
        sel.select(self.beat)
    File "python3.8/lib/python3.8/selectors.py", line 323, in select
        r, w, _ = self._select(self._readers, self._writers, [], timeout)
    ValueError: filedescriptor out of range in select()

But using PollSelector it is not reproduced even using more workers

Most of platform supports PollSelector but using DefaultSelector we can be sure
that even too old system are supported too

And using this High-level library will allow to use the future new improvements

e.g. Epoll has better performance improvements

More info about:
 - https://devarea.com/linux-io-multiplexing-select-vs-poll-vs-epoll
 - https://github.com/redis/redis-py/issues/486

Forcing the current SelectSelector:
 - ![Screen Shot 2022-02-15 at 14 28 00](https://user-images.githubusercontent.com/6644187/154143826-ad99ba18-978d-4435-9735-ed1425ef055d.png)

Using the DefaultSelector:
 - ![Screen Shot 2022-02-15 at 14 32 47](https://user-images.githubusercontent.com/6644187/154144432-bf67e24a-2732-46cb-85ee-f277a1a53f12.png)

Even if you don't need too many workers because of horizontal autoscaling it will have better performance and auto-evolution in the future python versions or systems

If you like this PR for v14.0 you can check the following PR:
 - https://github.com/Vauxoo/odoo/pull/477