@moylop260 2022-02-15
@odony

This is the change I told you a few months ago in twitter

cc @xmo-odoo