@Julien00859 2022-02-18
The ci/styles is complaining that you don't use the `mask` variable from your loop. The way to get rid of this is to prefix the `mask` variable with an understore: `for ..., _mask in ...`