@Julien00859 2022-02-18
I did a quick benchmark on my machine. The values are not be be taken by themselves, the important bit is the difference between master and your branch. Empty clean databases, only the test_http module installed. Server running without crons and with 16 http workers (the maximum for my machine). Tested with `ab` from `apache2-utils`, sending 24000 requests to /test_http/greeting, 24 parallels requests at a time, with a session set.

    ~ $ set session_id (curl -v odoo.localhost/test_http/greeting 2>&1| grep session_id | cut -d '=' -f2 | cut -d ';' -f1); echo $session_id
    d875b005768a01a6f5876ab533806d4a0e60a46c
    ~ $ ab -n 24000 -c 24 -C session_id=$session_id 127.0.0.1:8069/test_http/greeting-none

Master:

    This is ApacheBench, Version 2.3 <$Revision: 1843412 $>
    Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
    Licensed to The Apache Software Foundation, http://www.apache.org/

    Benchmarking 127.0.0.1 (be patient)


    Server Software:        Werkzeug/0.16.1
    Server Hostname:        127.0.0.1
    Server Port:            8069

    Document Path:          /test_http/greeting-none
    Document Length:        9 bytes

    Concurrency Level:      24
    Time taken for tests:   19.844 seconds
    Complete requests:      24000
    Failed requests:        0
    Total transferred:      7224000 bytes
    HTML transferred:       216000 bytes
    Requests per second:    1209.44 [#/sec] (mean)
    Time per request:       19.844 [ms] (mean)
    Time per request:       0.827 [ms] (mean, across all concurrent requests)
    Transfer rate:          355.51 [Kbytes/sec] received

    Connection Times (ms)
                  min  mean[+/-sd] median   max
    Connect:        0    0   0.4      0      15
    Processing:     3   19 113.1     12    2836
    Waiting:        3   19 113.1     12    2835
    Total:          3   19 113.2     12    2836

    Percentage of the requests served within a certain time (ms)
      50%     12
      66%     14
      75%     15
      80%     15
      90%     18
      95%     21
      98%     28
      99%     51
     100%   2836 (longest request)


This branch:

    This is ApacheBench, Version 2.3 <$Revision: 1843412 $>
    Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
    Licensed to The Apache Software Foundation, http://www.apache.org/

    Benchmarking 127.0.0.1 (be patient)


    Server Software:        Werkzeug/0.16.1
    Server Hostname:        127.0.0.1
    Server Port:            8069

    Document Path:          /test_http/greeting-none
    Document Length:        9 bytes

    Concurrency Level:      24
    Time taken for tests:   20.762 seconds
    Complete requests:      24000
    Failed requests:        0
    Total transferred:      7224000 bytes
    HTML transferred:       216000 bytes
    Requests per second:    1155.97 [#/sec] (mean)
    Time per request:       20.762 [ms] (mean)
    Time per request:       0.865 [ms] (mean, across all concurrent requests)
    Transfer rate:          339.79 [Kbytes/sec] received

    Connection Times (ms)
                  min  mean[+/-sd] median   max
    Connect:        0    0   0.8      0      30
    Processing:     3   20 114.9     13    2706
    Waiting:        3   19 114.8     12    2706
    Total:          3   20 115.0     13    2706

    Percentage of the requests served within a certain time (ms)
      50%     13
      66%     14
      75%     15
      80%     16
      90%     19
      95%     23
      98%     31
      99%     51
     100%   2706 (longest request)

The total time, 19s844 in master vs 20s762 in your branch bothers me. There is a 4.5% perf loose. Would you mind running a benchmark on your side too?