@moylop260 2022-02-18
@Julien00859 

I think if you run the same test using even the same branch you will have different results in the `Time taken for tests`

Try running almost 3 times the same test and you will see it