@moylop260 2022-02-21
@Julien00859 

I ran the same tests too many times and I saw the same results than you ~4.5% perf loose

I just realized that other kind of Selectors is using timeout sleep in ms instead of s

So, it needs a extra compute to convert 
 - `math.ceil(timeout * 1e3)`

and the selectors library is validating the input and parsing the output in order to be all them compatible with all Selectors using the same code

So the time for each request using Select directly (without high-level library) 0.467ms
vs EPoll with high-level library 0.447ms
Diff 0.020ms

Let me know if we should develop our own library using directly ms instead of seconds in order to avoid recomputing this value here:
 - https://github.com/odoo/odoo/blob/69e31db6ac90874ed6329599fca8f09709efd119/odoo/service/server.py#L700

I mean,

```patch
self.beat = math.ceil(4 * 1e3) 
```

and add code to validate if the system is compatible with `Poll` if not using `Select` by default and using the original class directly without extra validations

Using this way the time could be similar using directly the original class but the code added will be even more (since that it could be considered as a custom faster high-level library)

but it will fix the filedescriptor error when you are using workers>=255

Let me know what should we do in order to change it, please

- Compatibility with current 5 selectors using high-level library ~4.5% perf loose
- Compatibility with only 2 selectors (Select & Poll) using directly the original classes with extra-validations without perf loose