@Julien00859 2022-02-21
Hello @moylop260 it is not up to me to decide, I merely did a quick code review and asserted the performances. Your last comment confuses me more than it helps, what new library? From the python documentation, the timeout must be specified in seconds, I don't understand why you say sometimes it takes ms.

https://docs.python.org/3/library/selectors.html#selectors.BaseSelector.select

> If timeout > 0, this specifies the maximum wait time, in seconds.

The actual code is clean, I would be against we develop yet another abstraction layer around selector or that we reject kqueue/epoll/iocp. It is 4.5% slower, ok, it is what it is. Now we should wait and see what Olivier and Xavier have to say about this branch.