@moylop260 2022-02-21
> I don't understand why you say sometimes it takes ms.

Yes, we are using a High-level library to avoid using each particular case of the selector in a low-level

Odoo is using `Select` in a low-level and it is using the low-level timeout is in seconds for `Select` but milliseconds for `Poll`

So, someone needs to transform the seconds used by the high-level library to miliseconds

It is in the following line:
 - https://github.com/python/cpython/blob/c60414de7cefd092643ba200c2c045da1569c391/Lib/selectors.py#L410-L412

because low-level `Poll` is using milliseconds but we are sending seconds

So, we can implement low-level `Poll` but using the timeout in milliseconds directly in the variables in order to avoid transforming `self.beat = math.ceil(4 * 1e3)`

Another extra processing is transforming the output to match with all selectors in a standard way here:
 - https://github.com/python/cpython/blob/c60414de7cefd092643ba200c2c045da1569c391/Lib/selectors.py#L418-L427

So, it is spending the `0.020ms`