@moylop260 2022-08-09
Superseded by https://github.com/odoo/odoo/pull/97831